clean unnecessary comments and import, remove unused methods, rename task class, delete trailing white spaces and fix tabs.
